cc2: Use unsigned for the alignment
The biggest alignment of a C type fits in an integer for sure.

+++ src/cmd/scc-cc/cc2/cc2.h
@@ -154,7 +154,7 @@ typedef struct inst Inst;
 
 struct type {
 	unsigned long size;
-	unsigned long align;
+	unsigned align;
 	unsigned short id;
 	short flags;
 };
